Cache gateware build using `coppermind`
This makes use of a buildkite plugin named `coppermind` which checks the
treehash of a set of inputs, and if a build has been run previously with
that treehash, it just downloads the previously-built gateware, rather
than rebuilding it.  In order to do this, we must provide S3
credentials, so we reach for our trusty friend `cryptic` to get that
done.