fixup streaming behavior on overflow
this attempts to make sure that the streams amongst multiple devices
maintain the same DMA counters. Early behavior would try to
evacuate buffers on an overflow, which would cause a plurality
of devices to come out of sync. One approach is to try to "resync".
Hoever it is simpler to accept this corruption and return a
Overflow indication such that it can be marked in the stream.

We also remove the "partial buffer" logic since that should never happen...
we will always recieve a complete MTU from teh kernel.